Refactor the `chip` and `flash_target` modules into a new `targets` module
