+++ espflash/src/targets/esp32c2.rs
@@ -1,0 +1,139 @@
+use std::{collections::HashMap, ops::Range};
+
+use esp_idf_part::PartitionTable;
+
+use super::{Chip, Esp32Params, Esp32Target, FlashTarget, ReadEFuse, SpiRegisters, Target};
+use crate::{
+    connection::Connection,
+    elf::FirmwareImage,
+    error::Error,
+    flasher::{FlashFrequency, FlashMode, FlashSize, SpiAttachParams},
+    image_format::{Esp32BootloaderFormat, Esp32DirectBootFormat, ImageFormat, ImageFormatId},
+};
+
+pub(crate) const CHIP_DETECT_MAGIC_VALUES: &[u32] = &[
+    0x6f51_306f, // ECO0
+    0x7c41_a06f, // ECO1
+];
+
+pub(crate) const FLASH_RANGES: &[Range<u32>] = &[
+    0x4200_0000..0x4240_0000, // IROM
+    0x3c00_0000..0x3c40_0000, // DROM
+];
+
+pub(crate) const PARAMS: Esp32Params = Esp32Params::new(
+    0x0,
+    0x1_0000,
+    0x1f_0000,
+    12,
+    include_bytes!("../../resources/bootloaders/esp32c2-bootloader.bin"),
+);
+
+pub struct Esp32c2;
+
+impl Esp32c2 {
+    pub fn has_magic_value(value: u32) -> bool {
+        CHIP_DETECT_MAGIC_VALUES.contains(&value)
+    }
+}
+
+impl ReadEFuse for Esp32c2 {
+    fn efuse_reg(&self) -> u32 {
+        0x6000_8800
+    }
+}
+
+impl Target for Esp32c2 {
+    fn addr_is_flash(&self, addr: u32) -> bool {
+        FLASH_RANGES.iter().any(|range| range.contains(&addr))
+    }
+
+    fn chip_features(&self, _connection: &mut Connection) -> Result<Vec<&str>, Error> {
+        Ok(vec!["WiFi"])
+    }
+
+    fn chip_revision(&self, connection: &mut Connection) -> Result<Option<u32>, Error> {
+        let block1_addr = self.efuse_reg() + 0x44;
+        let num_word = 3;
+        let pos = 18;
+
+        let value = connection.read_reg(block1_addr + (num_word * 0x4))?;
+        let value = (value & (0x7 << pos)) >> pos;
+
+        Ok(Some(value))
+    }
+
+    fn crystal_freq(&self, _connection: &mut Connection) -> Result<u32, Error> {
+        // The ESP32-C2's XTAL has a fixed frequency of 40MHz.
+        Ok(40)
+    }
+
+    fn flash_frequency_encodings(&self) -> HashMap<FlashFrequency, u8> {
+        use FlashFrequency::*;
+
+        let encodings = [
+            (Flash15M, 0x2),
+            (Flash20M, 0x1),
+            (Flash30M, 0x0),
+            (Flash60M, 0xF),
+        ];
+
+        HashMap::from(encodings)
+    }
+
+    fn flash_target(&self, spi_params: SpiAttachParams, use_stub: bool) -> Box<dyn FlashTarget> {
+        Box::new(Esp32Target::new(Chip::Esp32c2, spi_params, use_stub))
+    }
+
+    fn get_flash_image<'a>(
+        &self,
+        image: &'a dyn FirmwareImage<'a>,
+        bootloader: Option<Vec<u8>>,
+        partition_table: Option<PartitionTable>,
+        image_format: Option<ImageFormatId>,
+        _chip_revision: Option<u32>,
+        flash_mode: Option<FlashMode>,
+        flash_size: Option<FlashSize>,
+        flash_freq: Option<FlashFrequency>,
+    ) -> Result<Box<dyn ImageFormat<'a> + 'a>, Error> {
+        let image_format = image_format.unwrap_or(ImageFormatId::Bootloader);
+
+        match image_format {
+            ImageFormatId::Bootloader => Ok(Box::new(Esp32BootloaderFormat::new(
+                image,
+                Chip::Esp32c2,
+                PARAMS,
+                partition_table,
+                bootloader,
+                flash_mode,
+                flash_size,
+                flash_freq,
+            )?)),
+            ImageFormatId::DirectBoot => Ok(Box::new(Esp32DirectBootFormat::new(image, 0)?)),
+        }
+    }
+
+    fn spi_registers(&self) -> SpiRegisters {
+        SpiRegisters {
+            base: 0x6000_2000,
+            usr_offset: 0x18,
+            usr1_offset: 0x1C,
+            usr2_offset: 0x20,
+            w0_offset: 0x58,
+            mosi_length_offset: Some(0x24),
+            miso_length_offset: Some(0x28),
+        }
+    }
+
+    fn supported_image_formats(&self) -> &[ImageFormatId] {
+        &[ImageFormatId::Bootloader, ImageFormatId::DirectBoot]
+    }
+
+    fn supported_targets(&self) -> &[&str] {
+        &[
+            "riscv32imac-unknown-none-elf",
+            "riscv32imc-esp-espidf",
+            "riscv32imc-unknown-none-elf",
+        ]
+    }
+}
